% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPar.R
\name{calcPar}
\alias{calcPar}
\title{Endogenous switching regression}
\usage{
calcPar(Results)
}
\arguments{
\item{Results}{An maxLik object estimated from endoSwitch}
}
\value{
A matrix that provides estimates, standard errors, and z values.
}
\description{
This R function gives estimates of sigma and rho from an estimated model
}
\examples{
data(ImpactData)
ManDepVar <- 'Output'
SelDepVar <- 'CA'
ManCovVar <- c('Age')
SelCovVar <- c('Age', 'Perception')
Results <- endoSwitch(ImpactData, ManDepVar, SelDepVar, ManCovVar, SelCovVar)
summary(Results)

calcPar(Results)

}
