% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endoSwitch2Stage.R
\name{endoSwitch2Stage}
\alias{endoSwitch2Stage}
\title{Endogenous switching regression}
\usage{
endoSwitch2Stage(RegData, OutcomeDep, SelectDep, OutcomeCov, SelectCov)
}
\arguments{
\item{RegData}{Data for doing the regression analysis}

\item{OutcomeDep}{Dependent variable in the main model}

\item{SelectDep}{Dependent variable in the selection model. This should be binary (0 or 1).}

\item{OutcomeCov}{Independent variables in the main model.}

\item{SelectCov}{Independent variable in the selection model}
}
\value{
A list containing regression results.
}
\description{
This function estimates the endogenous switching regression model through two-stage regression.
The estimation procedures are described in Heckman et al. (2001). The \code{endoSwitch} function is
preferred than this function for estimation.
}
\details{
The first stage uses a probit model to estimate the selection equation. The second stage uses OLS, plus the inverse mills ratios to estimate the main equations.
}
\examples{
data(ImpactData)
OutcomeDep <- 'Output'
SelectDep <- 'CA'
OutcomeCov <- c('Age')
SelectCov <- c('Age', 'Perception')
Results <- endoSwitch2Stage(ImpactData, OutcomeDep, SelectDep, OutcomeCov, SelectCov)
# First stage regression results
summary(Results$FirstStageReg)
# Second stage regression results -- non-adopter
summary(Results$SecondStageReg.0)
# Second stage regression results -- adopter
summary(Results$SecondStageReg.1)
}
\references{
Heckman, Tobias, and Vytlacil (2001). Four parameters of interest in the evaluation of
social programs. \emph{Southern Economic Journal}.
}
