% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImpactData.R
\docType{data}
\name{ImpactData}
\alias{ImpactData}
\title{Data on conservation agriculture}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 408 rows and 31 columns.}
\source{
\href{https://onlinelibrary.wiley.com/doi/abs/10.1111/agec.12269}{Online}
}
\usage{
data(ImpactData)
}
\description{
Data on conservation agriculture
}
\examples{
data(ImpactData)
summary(ImpactData)
\donttest{iplotCurves(phe, times)}
}
\references{
Abdulai (2016) Agricultural Economics 47:729-741
(\href{https://onlinelibrary.wiley.com/doi/abs/10.1111/agec.12269}{PubMed})
}
\keyword{datasets}
