% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.endoSwitch.R
\name{summary.endoSwitch}
\alias{summary.endoSwitch}
\title{Summarize the endogenous switching regression results.}
\usage{
\method{summary}{endoSwitch}(object, ...)
}
\arguments{
\item{object}{Estimated endogenous switching regression model.}

\item{...}{Other elements.}
}
\value{
A list containing the key regression results.
}
\description{
Summarize the endogenous switching regression results.
}
