% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentEffect.R
\name{treatmentEffect}
\alias{treatmentEffect}
\title{Endogenous switching regression}
\usage{
treatmentEffect(
  Results,
  RegData,
  OutcomeDepVar,
  SelectDepVar,
  OutcomeCovVar,
  SelectCovVar
)
}
\arguments{
\item{Results}{Estimated endogenous switching regression model}

\item{RegData}{Data for doing the regression analysis}

\item{OutcomeDepVar}{Dependent variable in the main model}

\item{SelectDepVar}{Dependent variable in the Selection model. This should be binary (0 or 1).}

\item{OutcomeCovVar}{Independent variables in the main model.}

\item{SelectCovVar}{Independent variable in the Selection model}
}
\value{
A matrix that reports the calculated average treatment effects.
}
\description{
This function calculates average treatment effects from an estimated endogenous switching regression model,
including average treatment effects, average treatment effects on the treated, average treatment effects on
the untreated.
}
\references{
Abdulai (2016) Impact of conservation agriculture technology on
household welfare in Zambia. \emph{Agricultural Economics} 47:729-741
(\href{https://onlinelibrary.wiley.com/doi/abs/10.1111/agec.12269}{link here})
}
