% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPar.R
\name{calcPar}
\alias{calcPar}
\title{Endogenous switching regression}
\usage{
calcPar(Results)
}
\arguments{
\item{Results}{A maxLik object that is estimated by the \code{endoSwitch} function.}
}
\value{
A matrix that reports the estimates of original distribution parameters. sigma is standard deviation,
and rho is correlation coefficient.
}
\description{
This R function provides estimates of original distribution parameters (sigma and rho) from the estimates of
transformed distribution parameters using the delta method.
}
