% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endoSwitch.R
\name{endoSwitch}
\alias{endoSwitch}
\title{Endogenous switching regression}
\usage{
endoSwitch(
  RegData,
  ManDepVar,
  SelDepVar,
  ManCovVar,
  SelCovVar,
  method = "BFGS",
  start,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{RegData}{Data for doing the regression analysis}

\item{ManDepVar}{Dependent variable in the main model}

\item{SelDepVar}{Dependent variable in the selection model. This should be binary (0 or 1).}

\item{ManCovVar}{Independent variables in the main model.}

\item{SelCovVar}{Independent variable in the selection model}

\item{method}{Maximization method.}

\item{start}{Numeric vector used as initial value of parameters for maximization purpose (including the constants)}

\item{verbose}{Print the log likelihood values during the optimization? TRUE or FALSE}

\item{...}{Other parameters to be passed to the selected maximization routine.}
}
\value{
An object of class 'maxLik". The estimates include parameters in the selection model,
parameters in the main models in two different regimes, and the associated distributional parameters.
}
\description{
This is the main interface for the endoSwitch package, and the function performs Maximum Likelihood
estimation of an endogenous switching regression.
}
\details{
This R function estimates parameters in an endogenous switching regression model by maximizing the
joint likelihood function, and the formula for calculating the likilihood function is from
Lokshin and Sajaia (2004).

The function needs starting values to run. Assume that you have M variables (including the constant) in the selection model,
and N variables (including the constant) in the main model.
Then you need (M + 2*N + 4) starting values. The first M values are for the variables in the
selection model (last for the constant), the following N values for the main model with SelDepVar = 0,
and the following N values for the main model with SelDepVar = 1.
The last four values are: sigma in main model with SelDepVar = 0, sigma in main model with SelDepVar = 1,
rho in main model with SelDepVar = 0, rho in main model with SelDepVar = 1.
}
\examples{
data(ImpactData)
ManDepVar <- 'Output(ton)'
SelDepVar <- 'CA'
ManCovVar <- c('Age (years)')
SelCovVar <- c('Age (years)', 'Perception (1,0)')
Results <- endoSwitch(ImpactData, ManDepVar, SelDepVar, ManCovVar, SelCovVar,
start = c(-0.007, 2.15, -0.2, 0.0007, 1.6, 0.0058, 2.6, 0.5, 0.5, -0.1, 0.2))
summary(Results)
}
