% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endoSwitch2Stage.R
\name{endoSwitch2Stage}
\alias{endoSwitch2Stage}
\title{Endogenous switching regression}
\usage{
endoSwitch2Stage(RegData, ManDepVar, SelDepVar, ManCovVar, SelCovVar)
}
\arguments{
\item{RegData}{Data for doing the regression analysis}

\item{ManDepVar}{Dependent variable in the main model}

\item{SelDepVar}{Dependent variable in the selection model. This should be binary (0 or 1).}

\item{ManCovVar}{Independent variables in the main model.}

\item{SelCovVar}{Independent variable in the selection model}
}
\value{
A list containing regression results.
}
\description{
This function estimates the endogenous switching regression model through two-stage regression following Heckman et al. (2001).
The estimation procedures are described in
}
\details{
The first stage uses a probit model to estimate the selection equation. The second stage uses OLS to estimate the main equations.
}
\examples{
data(ImpactData)
ManDepVar <- 'Output'
SelDepVar <- 'CA'
ManCovVar <- c('Age')
SelCovVar <- c('Age', 'Perception')
Results <- endoSwitch2Stage(ImpactData, ManDepVar, SelDepVar, ManCovVar, SelCovVar)
# First stage regression results
summary(Results$FirstStageReg)
# Second stage regression results -- non-adopter
summary(Results$SecondStageReg_NoAdopt)
# Second stage regression results -- adopter
summary(Results$SecondStageReg_Adopt)
}
