% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImpactData.R
\docType{data}
\name{ImpactData}
\alias{ImpactData}
\title{A dataset on adoption of conservation agriculture in Zambia}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 408 rows and 31 columns.}
\source{
Abdulai (2016) Impact of conservation agriculture technology on
household welfare in Zambia. \emph{Agricultural Economics} 47:729-741
(\href{https://onlinelibrary.wiley.com/doi/abs/10.1111/agec.12269}{link here})
}
\usage{
data(ImpactData)
}
\description{
A dataset on adoption of conservation agriculture in Zambia
}
\examples{
data(ImpactData)
summary(ImpactData)
}
\references{
Abdulai (2016) Impact of conservation agriculture technology on
household welfare in Zambia. \emph{Agricultural Economics} 47:729-741
(\href{https://onlinelibrary.wiley.com/doi/abs/10.1111/agec.12269}{link here})
}
\keyword{datasets}
