% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImpactData.R
\docType{data}
\name{ImpactData}
\alias{ImpactData}
\title{A dataset on adoption of conservation agriculture in Zambia}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 408 rows and 31 columns.}
\source{
Abdulai, Abdul Nafeo. “Impact of Conservation Agriculture Technology on Household Welfare in Zambia.”
Agricultural Economics 47, no. 6 (2016): 729–41. https://doi.org/10.1111/agec.12269.
}
\usage{
data(ImpactData)
}
\description{
A dataset on adoption of conservation agriculture in Zambia
}
\examples{
data(ImpactData)
summary(ImpactData)
}
\references{
Abdulai, Abdul Nafeo. “Impact of Conservation Agriculture Technology on Household Welfare in Zambia.”
Agricultural Economics 47, no. 6 (2016): 729–41. https://doi.org/10.1111/agec.12269.
}
\keyword{datasets}
