% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentEffect.R
\name{treatmentEffect}
\alias{treatmentEffect}
\title{Endogenous switching regression}
\usage{
treatmentEffect(Results, RegData, ManDepVar, SelDepVar, ManCovVar, SelCovVar)
}
\arguments{
\item{Results}{Estimated endogenous switching regression model}

\item{RegData}{Data for doing the regression analysis}

\item{ManDepVar}{Dependent variable in the main model}

\item{SelDepVar}{Dependent variable in the selection model. This should be binary (0 or 1).}

\item{ManCovVar}{Independent variables in the main model.}

\item{SelCovVar}{Independent variable in the selection model}
}
\value{
A matrix that reports the calculated average treatment effects.
}
\description{
This function calculates average treatment effects from an estimated endogenous switching regression model,
including average treatment effects, average treatment effects on the treated, average treatment effects on
the untreated.
}
\examples{
data(ImpactData)
ManDepVar <- 'Output'
SelDepVar <- 'CA'
ManCovVar <- c('Age')
SelCovVar <- c('Age', 'Perception')
Results <- endoSwitch(ImpactData, ManDepVar, SelDepVar, ManCovVar, SelCovVar)
treatmentEffect(Results, ImpactData, ManDepVar, SelDepVar, ManCovVar, SelCovVar)
}
\references{
Abdulai (2016) Impact of conservation agriculture technology on
household welfare in Zambia. \emph{Agricultural Economics} 47:729-741
(\href{https://onlinelibrary.wiley.com/doi/abs/10.1111/agec.12269}{link here})
}
