% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentEffect.R
\name{treatmentEffect}
\alias{treatmentEffect}
\title{Endogenous switching regression}
\usage{
treatmentEffect(
  Results,
  data,
  OutcomeDep,
  SelectDep,
  OutcomeCov,
  SelectCov,
  treatEffect
)
}
\arguments{
\item{Results}{Estimated endogenous switching regression model}

\item{data}{an data frame. Data for running the regression analysis.}

\item{OutcomeDep}{character. Dependent variable in the outcome equation.}

\item{SelectDep}{character. Dependent variable in the Selection model. The iable must be binary (0 or 1).}

\item{OutcomeCov}{character vector. Covariates in the outcome equation.}

\item{SelectCov}{character vector. Covariates in the selection equation.}

\item{treatEffect}{TRUE/FALSE. Show average treatment effects or not.}
}
\value{
A matrix that reports the calculated average treatment effects.
}
\description{
This function calculates average treatment effects from an estimated endogenous switching regression model,
including average treatment effects, average treatment effects on the treated, average treatment effects on
the untreated.
}
\references{
Abdulai (2016) Impact of conservation agriculture technology on
household welfare in Zambia. \emph{Agricultural Economics} 47:729-741
(\href{https://onlinelibrary.wiley.com/doi/abs/10.1111/agec.12269}{link here})
}
